refactor: isolate project index parser facade
