+++ src/plugin/src/identifier-case/project-index-bootstrap.js
@@ -14,9 +14,13 @@ import {
 } from "../../../shared/object-utils.js";
 import {
     findProjectRoot,
-    createProjectIndexCoordinator,
-    getProjectIndexParserOverride
+    createProjectIndexCoordinator
 } from "../project-index/index.js";
+import { getProjectIndexParserOverride } from "../project-index/parser-override.js";
+import {
+    createProjectIndexBuildOptions,
+    createProjectIndexDescriptor
+} from "../project-index/bootstrap-descriptor.js";
 import { clampConcurrency } from "../project-index/concurrency.js";
 
 const PROJECT_INDEX_CACHE_MAX_BYTES_INTERNAL_OPTION_NAME =
@@ -374,51 +378,6 @@ function resolveProjectIndexCoordinator(
     return { coordinator, dispose };
 }
 
-function createProjectIndexBuildOptions(
-    options,
-    { projectIndexConcurrency, parserOverride }
-) {
-    const buildOptions = {
-        logger: options?.logger ?? null,
-        logMetrics: options?.logIdentifierCaseMetrics === true
-    };
-
-    withDefinedValue(projectIndexConcurrency, (value) => {
-        buildOptions.concurrency = {
-            gml: value,
-            gmlParsing: value
-        };
-    });
-
-    if (parserOverride) {
-        if (parserOverride.facade) {
-            buildOptions.gmlParserFacade = parserOverride.facade;
-        }
-        buildOptions.parseGml = parserOverride.parse;
-    }
-
-    return buildOptions;
-}
-
-function createProjectIndexDescriptor(
-    options,
-    { projectRoot, cacheMaxSizeBytes, buildOptions }
-) {
-    const descriptor = {
-        projectRoot,
-        cacheFilePath: options?.identifierCaseProjectIndexCachePath ?? null,
-        formatterVersion: getFormatterVersion(options) ?? undefined,
-        pluginVersion: getPluginVersion(options) ?? undefined,
-        buildOptions
-    };
-
-    withDefinedValue(cacheMaxSizeBytes, (value) => {
-        descriptor.maxSizeBytes = value;
-    });
-
-    return descriptor;
-}
-
 function finalizeBootstrapSuccess(
     options,
     ready,
@@ -500,14 +459,19 @@ export async function bootstrapProjectIndex(options = {}, storeOption) {
     });
 
     const parserOverride = getProjectIndexParserOverride(options);
-    const buildOptions = createProjectIndexBuildOptions(options, {
+    const buildOptions = createProjectIndexBuildOptions({
+        logger: options?.logger ?? null,
+        logMetrics: options?.logIdentifierCaseMetrics === true,
         projectIndexConcurrency,
         parserOverride
     });
 
-    const descriptor = createProjectIndexDescriptor(options, {
+    const descriptor = createProjectIndexDescriptor({
         projectRoot,
         cacheMaxSizeBytes,
+        cacheFilePath: options?.identifierCaseProjectIndexCachePath ?? null,
+        formatterVersion: getFormatterVersion(options) ?? undefined,
+        pluginVersion: getPluginVersion(options) ?? undefined,
         buildOptions
     });
