+++ src/plugin/src/project-index/index.js
@@ -21,7 +21,10 @@ import {
     buildLocationKey,
     buildFileLocationKey
 } from "../../../shared/location-keys.js";
-import { getDefaultProjectIndexParser } from "./gml-parser-facade.js";
+import {
+    
+    resolveProjectIndexParser
+} from "./parser-override.js";
 import { clampConcurrency } from "./concurrency.js";
 import {
     PROJECT_MANIFEST_EXTENSION,
@@ -51,14 +54,6 @@ import {
 import { loadBuiltInIdentifiers } from "./built-in-identifiers.js";
 import { createProjectIndexCoordinatorFactory } from "./coordinator.js";
 
-const defaultProjectIndexParser = getDefaultProjectIndexParser();
-
-const PARSER_FACADE_OPTION_KEYS = [
-    "identifierCaseProjectIndexParserFacade",
-    "gmlParserFacade",
-    "parserFacade"
-];
-
 /**
  * Create shallow clones of common entry collections stored on project index
  * records (for example declaration/reference lists). Guarding against
@@ -69,32 +64,6 @@ function cloneEntryCollections(entry, ...keys) {
     const source = isObjectLike(entry) ? entry : {};
     return Object.fromEntries(
         keys.map((key) => [key, cloneObjectEntries(source[key])])
-    );
-}
-
-function getProjectIndexParserOverride(options) {
-    if (!isObjectLike(options)) {
-        return null;
-    }
-
-    for (const key of PARSER_FACADE_OPTION_KEYS) {
-        const facade = options[key];
-        if (typeof facade?.parse === "function") {
-            return {
-                facade,
-                parse: facade.parse.bind(facade)
-            };
-        }
-    }
-
-    const parse = options.parseGml;
-    return typeof parse === "function" ? { facade: null, parse } : null;
-}
-
-function resolveProjectIndexParser(options) {
-    return (
-        getProjectIndexParserOverride(options)?.parse ??
-        defaultProjectIndexParser
     );
 }
 
@@ -2385,6 +2354,7 @@ export async function buildProjectIndex(
     });
 }
 export { defaultFsFacade } from "./fs-facade.js";
-export { getProjectIndexParserOverride };
+
 export { ProjectFileCategory };
 export { __loadBuiltInIdentifiersForTests } from "./built-in-identifiers.js";
+export {getProjectIndexParserOverride} from "./parser-override.js";
